# [algo, perf] feat: Vectorize GRPO Advantage Estimator - 13～26x Speedup 