--- a/tests/utils/test_groupwise.py
+++ b/tests/utils/test_groupwise.py
@@ -1,10 +1,25 @@
-# -*- coding: utf-8 -*-
+# Copyright 2024 Bytedance Ltd. and/or its affiliates
+# Copyright 2023-2024 SGLang Team
+# Copyright 2025 ModelBest Inc. and/or its affiliates
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 import os
+
 os.environ.setdefault("VERL_FORCE_DEVICE", "cpu")  # ensure CPU for tests
 
 import numpy as np
-import torch
 import pytest
+import torch
 
 from verl.utils import as_torch_index, group_mean_std
 
@@ -47,7 +62,7 @@ def test_group_mean_std_simple():
     # singleton group -> std = 1.0
     assert mean_g[1].item() == 0.0
     assert std_g[1].item() == 1.0
-    assert pytest.approx(std_g[0].item(), rel=1e-6) == (2.0 ** 0.5)
+    assert pytest.approx(std_g[0].item(), rel=1e-6) == (2.0**0.5)
 
 
 def test_group_mean_std_empty():