--- a/verl/utils/__init__.py
+++ b/verl/utils/__init__.py
@@ -14,12 +14,12 @@
 
 from . import config, tokenizer
 from .config import omega_conf_to_dataclass, validate_config
-from .tokenizer import hf_processor, hf_tokenizer
 from .groupwise import as_torch_index, group_mean_std
+from .tokenizer import hf_processor, hf_tokenizer
 
 __all__ = (
-    tokenizer.__all__ 
-    + config.__all__ 
+    tokenizer.__all__
+    + config.__all__
     + ["hf_processor", "hf_tokenizer", "omega_conf_to_dataclass", "validate_config"]
     + ["as_torch_index", "group_mean_std"]
 )